[test] Add test_fuse_dynamic.py to unit test
